🗃️ refactor: change latitude and longitude type to decimal

+++ database/migrations/20240314014504-create-treasures.js
@@ -12,11 +12,11 @@ module.exports = {
         autoIncrement: false,
       },
       latitude: {
-        type: DataTypes.FLOAT(11, 8),
+        type: DataTypes.DECIMAL(10, 8),
         allowNull: false,
       },
       longitude: {
-        type: DataTypes.FLOAT(11, 8),
+        type: DataTypes.DECIMAL(11, 8),
         allowNull: false,
       },
       name: {
